move firebase mod to core
