+++ libs/server/common/src/core/firebase/index.ts
