+++ libs/server/common/src/core/firebase/logger.util.ts
