+++ libs/server/common/src/gamification/xp-database.service.ts
@@ -1,7 +1,7 @@
 import { Firestore } from 'firebase-admin/firestore';
 import * as logger from 'firebase-functions/logger';
 import { UserActivity } from '../activity';
-import { DatabaseInstance, getCurrentTimeAsISO } from '../firebase';
+import { DatabaseInstance, getCurrentTimeAsISO } from '../core/firebase';
 import { XpCalculationResponse, XpHistoryEntry } from './gamification-domain.model';
 import { calculateLevel } from './levels/level.utils';
