+++ libs/server/common/src/index.ts
@@ -1,7 +1,7 @@
 export * from './activity';
 export * from './core';
 export * from './event';
-export * from './firebase';
+export * from './core/firebase';
 export * from './gamification';
 export * from './user';
